Code Cleanup from VS 2019

+++ Library/ClassLibraryCSVUnitTest/SampleRecordEntryTests.cs
@@ -25,16 +25,16 @@ namespace CsvTools.Tests
       Assert.AreEqual(true, entry.ProvideEvidence);
 
       var entry1 = new SampleRecordEntry(100, "Error");
-      Assert.AreEqual((long)100, entry1.RecordNumber);
+      Assert.AreEqual(100, entry1.RecordNumber);
       Assert.AreEqual("Error", entry1.Error);
       Assert.AreEqual(true, entry1.ProvideEvidence);
 
       var entry2 = new SampleRecordEntry(1000, false);
-      Assert.AreEqual((long)1000, entry2.RecordNumber);
+      Assert.AreEqual(1000, entry2.RecordNumber);
       Assert.AreEqual(false, entry2.ProvideEvidence);
 
       var entry3 = new SampleRecordEntry(2000);
-      Assert.AreEqual((long)2000, entry3.RecordNumber);
+      Assert.AreEqual(2000, entry3.RecordNumber);
     }
 
 
@@ -43,7 +43,7 @@ namespace CsvTools.Tests
     {
       var entry1 = new SampleRecordEntry(100, "Error1");
       var entry2 = entry1.Clone();
-      Assert.AreEqual((long)100, entry2.RecordNumber);
+      Assert.AreEqual(100, entry2.RecordNumber);
       Assert.AreEqual("Error1", entry2.Error);
       Assert.IsTrue(entry2.ProvideEvidence);
     }
